Change module structure.
