+++ .idea/gradle.xml
@@ -19,6 +19,9 @@
             <option value="$PROJECT_DIR$/data/repository" />
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/home" />
+            <option value="$PROJECT_DIR$/feature/home/compose" />
+            <option value="$PROJECT_DIR$/feature/home/presentation" />
+            <option value="$PROJECT_DIR$/feature/home/usecase" />
             <option value="$PROJECT_DIR$/testUtil" />
             <option value="$PROJECT_DIR$/viewCommon" />
           </set>
