--- feature/home/build.gradle
@@ -1,9 +1,0 @@
-apply plugin: 'com.android.library'
-apply from: rootProject.file("gradle/view-common.gradle")
-
-dependencies {
-    implementation project(':common')
-    implementation project(':viewCommon')
-    implementation project(':data:repository')
-    testImplementation project(':testUtil')
-}
