+++ feature/home/presentation/build.gradle
@@ -1,0 +1,10 @@
+apply plugin: 'com.android.library'
+apply from: rootProject.file("gradle/view-common.gradle")
+
+dependencies {
+    implementation project(':common')
+    implementation project(':viewCommon')
+    implementation project(':feature:home:compose')
+    implementation project(':feature:home:usecase')
+    testImplementation project(':testUtil')
+}
