+++ feature/home/usecase/build.gradle
@@ -1,0 +1,7 @@
+apply plugin: 'com.android.library'
+apply from: rootProject.file("gradle/common.gradle")
+
+dependencies {
+    implementation project(':common')
+    implementation project(':data:repository')
+}
