+++ settings.gradle
@@ -12,6 +12,8 @@ include ':common'
 include ':data:remote'
 include ':data:local'
 include ':data:repository'
-include ':feature:home'
 include ':viewCommon'
 include ':testUtil'
+include ':feature:home:compose'
+include ':feature:home:presentation'
+include ':feature:home:usecase'
